Remove updateSubscription from APIConsumer level and added that to the super class
